more checks to stop PHP warnings

+++ pro/includes/class-foogallery-pro-datasource-realmedialibrary.php
@@ -90,9 +90,11 @@ if ( ! class_exists( 'FooGallery_Pro_Datasource_RealMediaLibrary' ) ) {
 			$previous_datasource_value = get_post_meta( $foogallery_id, FOOGALLERY_META_DATASOURCE_VALUE, true );
 
 			if ( is_array( $previous_datasource_value ) && array_key_exists( 'value', $previous_datasource_value ) ) {
-				$fid                 = $previous_datasource_value['value'];
-				$cache_post_meta_key = FOOGALLERY_META_DATASOURCE_CACHED_ATTACHMENTS . '_' . $fid;
-				delete_post_meta( $foogallery_id, $cache_post_meta_key );
+				$fid = $previous_datasource_value['value'];
+				if ( is_string( $fid ) ) {
+					$cache_post_meta_key = FOOGALLERY_META_DATASOURCE_CACHED_ATTACHMENTS . '_' . $fid;
+					delete_post_meta( $foogallery_id, $cache_post_meta_key );
+				}
 			}
 		}
